@@ -21,7 +21,7 @@ use Tmdb\Common\ParameterBag;
 use Tmdb\HttpClient\HttpClient;
 use Tmdb\HttpClient\Request;
 
-abstract class TestCase extends \PHPUnit_Framework_TestCase
+abstract class TestCase extends \PHPUnit\Framework\TestCase
 {
     protected $adapter = null;
 
