@@ -19,7 +19,7 @@ class CertificationFactoryTest extends TestCase
 {
     private $data;
 
-    public function setUp()
+    public function setUp() :void
     {
         $this->data = $this->loadByFile('certification/get.json');
     }
