@@ -22,7 +22,7 @@ class CompanyFactoryTest extends TestCase
 
     private $data;
 
-    public function setUp()
+    public function setUp() :void
     {
         $this->data = $this->loadByFile('company/get.json');
     }
