@@ -18,7 +18,7 @@ class KeywordFactoryTest extends TestCase
 {
     private $data;
 
-    public function setUp()
+    public function setUp() :void
     {
         $this->data = $this->loadByFile('keywords/get.json');
     }
