@@ -24,7 +24,7 @@ class ListFactoryTest extends TestCase
      */
     private $lists;
 
-    public function setUp()
+    public function setUp() :void
     {
         $this->data = $this->loadByFile('lists/get.json');
 
