@@ -31,7 +31,7 @@ class AcceptJsonHeaderPluginTest extends TestCase
 
         $eventDispatcher = new EventDispatcher();
         $eventDispatcher->addSubscriber(new AcceptJsonHeaderPlugin());
-        $eventDispatcher->dispatch(TmdbEvents::BEFORE_REQUEST, $event);
+        $eventDispatcher->dispatch($event, TmdbEvents::BEFORE_REQUEST);
 
         $this->assertEquals('application/json', (string) $event->getRequest()->getHeaders()->get('Accept'));
     }
